add starter for the react-cinema lab
