+++ client/src/App.jsx
@@ -1,10 +1,10 @@
-import React, { Component } from "react";
-import Home from "./components/Home";
+import React, { Component } from 'react';
+import Home from './components/Home';
 
 class App extends Component {
   render() {
     return (
-      <div className="App">
+      <div className='App'>
         <Home />
       </div>
     );
