+++ client/src/components/Home.jsx
@@ -1,10 +1,10 @@
-import React, { Component } from "react";
-import axios from "axios";
+import React, { Component } from 'react';
+import axios from 'axios';
 
 export default class Home extends Component {
   async callMyBackend() {
     // this is an example call to our backend
-    let responseFromBackend = await axios.get("http://localhost:5000/");
+    let responseFromBackend = await axios.get('http://localhost:5000/');
     console.log(responseFromBackend);
   }
   render() {
