+++ routes/index.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const router = express.Router();
+
+/* GET home page */
+router.get('/', (req, res) => {
+  console.log('called');
+  res.json({ message: 'index' });
+});
+
+module.exports = router;
