--- starter_code/.gitkeep
@@ -1,2 +1,0 @@
-# Dependency directory
-node_modules
