--- starter_code/app.js
@@ -1,39 +1,0 @@
-require("dotenv").config();
-
-const bodyParser = require("body-parser");
-const cookieParser = require("cookie-parser");
-const express = require("express");
-const mongoose = require("mongoose");
-const logger = require("morgan");
-const path = require("path");
-const cors = require("cors");
-
-mongoose
-  .connect("mongodb://localhost/starter-code", { useNewUrlParser: true })
-  .then((x) => {
-    console.log(
-      `Connected to Mongo! Database name: "${x.connections[0].name}"`
-    );
-  })
-  .catch((err) => {
-    console.error("Error connecting to mongo", err);
-  });
-
-const app_name = require("./package.json").name;
-const debug = require("debug")(
-  `${app_name}:${path.basename(__filename).split(".")[0]}`
-);
-
-const app = express();
-
-// Middleware Setup
-app.use(cors());
-app.use(logger("dev"));
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(cookieParser());
-
-const index = require("./routes/index");
-app.use("/", index);
-
-module.exports = app;
