--- starter_code/routes/index.js
@@ -1,10 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-/* GET home page */
-router.get("/", (req, res, next) => {
-  console.log("called");
-  res.json({ message: "index" });
-});
-
-module.exports = router;
